- first prototype!

+++ www/app/modules/Hub/templates/LinkEditInput.php
@@ -6,7 +6,9 @@
 		<label>
 			<span class="label">Title:</span>
 			<input type="text" name="title" size="50" maxlength="50" />
-			<span class="small">Max. 50 characters</span>
+			<span class="small">
+				E.g. Homepage, Download, Documentation, Demo, Screenshots. Max. 50 characters.
+			</span>
 		</label>
 
 		<label>
@@ -23,9 +25,11 @@
 		<label>
 			<span class="label">Priority:</span>
 			<select name="priority">
-				<option value="0">High - Official</option>
-				<option value="1" selected="selected">Normal - Community</option>
-				<option value="2">Low - Individual</option>
+				<option value="0">Very High - Official</option>
+				<option value="1">High</option>
+				<option value="2" selected="selected">Normal - Community</option>
+				<option value="3">Low - Individual</option>
+				<option value="4">Very Low - Additional Readings</option>
 			</select>
 			<span class="small">For the ranking on the project page.</span>
 		</label>
@@ -33,15 +37,13 @@
 	</fieldset>
 
 	<fieldset>
-
+<?php	if ($url_delete): ?>
 		<div class="alignright">
 			<a href="<?= $url_delete ?>">Delete</a>
 		</div>
+<?php	endif; ?>
 
-		<div class="buttons">
-			<input type="submit" value="Save Changes" class="submit" />
-			or <a href="<?= $url ?>">Cancel</a>
-		</div>
-
+		<input type="submit" value="Save Changes" class="submit" />
+		or <a href="<?= $url ?>">Cancel</a>
 	</fieldset>
 </form>
